+++ analysis/import/MetricGardenerImporter/src/main/kotlin/de/maibornwolff/codecharta/importer/metricgardenerimporter/ParserDialog.kt
@@ -9,15 +9,16 @@ class ParserDialog {
     companion object : ParserDialogInterface {
 
         override fun collectParserArgs(): List<String> {
-            val mgJsonAvailable: Boolean =
+            val isJsonFile: Boolean =
                 KInquirer.promptConfirm(
                     message = "Do you already have a MetricGardener json-File?",
-                    default = true
+                    default = false
                 )
 
             val inputFile = KInquirer.promptInput(
-                message = "What Project do you want to parse?",
-                hint = "path/to/my/project"
+                message = if (isJsonFile) "Which MetricGardener json-File do you want to import?"
+                          else "What Project do you want to parse?",
+                hint = if (isJsonFile) "path/to/metricgardener/file.json" else "path/to/my/project"
             )
 
             val outputFileName: String = KInquirer.promptInput(
@@ -32,7 +33,7 @@ class ParserDialog {
                 inputFile,
                 "--output-file=$outputFileName",
                 if (isCompressed) null else "--not-compressed",
-                if (mgJsonAvailable) null else "--with-mg-run"
+                if (isJsonFile) "--is-json-file" else null
             )
         }
     }